MIDI output

Basic support for MIDI output are provided.

Synth.midioutput — Function

midioutput(name::AbstractString = "") :: MIDIOutput

Identifies a MIDI output device with the given name as a substring and opens a stream to write data to it. You can call Base.close on the resultant MIDIOutput to close the stream.

Throws MIDIOutputDeviceNotFoundError if such a device does not exist.

Synth.send — Function

send(dev::MIDIOutput, msg::MIDIMsg)

Sends the MIDIMsg to the device immediately.

Synth.noteon — Function

noteon(chan::Int, note::Int, vel::Int) :: MIDIMsg
noteon(chan::Int, note::Int, vel::AbstractFloat) :: MIDIMsg

Constructs a NoteOn message. The AbstractFloat version will take a velocity value in the range 0.0 to 1.0 and convert it to the MIDI range. Note that for sustained instruments like violin, you'll have to also send a noteoff at an appropriate time or else the voices will accumulate and result in your system being unable to keep up.

See also noteoff

Synth.noteoff — Function

noteoff(chan::Int, note::Int, vel::Int = 0) :: MIDIMsg

Constructs a NoteOff message. See also noteon

Synth.progchange — Function

progchange(chan::Int, pgm::Int)

Send a "program change" for the given channel. See General MIDI if you're using a GM compatible synthesizer.

Synth.ctrlchange — Function

ctrlchange(chan::Int, control::Int, val::Int) :: MIDIMsg
ctrlchange(chan::Int, control::Int, val::AbstractFloat) :: MIDIMsg

Changes the value associated with the given control number. The AbstractFloat value is in the range 0.0-1.0 and will be rescaled to the MIDI range 0-127.

Synth.keypressure — Function

keypressure(chan::Int, note::Int, pressure::Int) :: MIDIMsg
keypressure(chan::Int, note::Int, pressure::AbstractFloat) :: MIDIMsg

The key pressure control after a note is turned on, before the note has conceptually ended. See also aftertouch. The AbstractFloat version will rescale the pressure value from the 0.0-1.0 range to the MIDI range of 0-127.

Synth.aftertouch — Function

aftertouch(chan::Int, note::Int, pressure::Int) :: MIDIMsg
aftertouch(chan::Int, note::Int, pressure::AbstractFloat) :: MIDIMsg

The key pressure control after a note has conceptually ended. See also aftertouch. The AbstractFloat version will rescale the pressure value from the 0.0-1.0 range to the MIDI range of 0-127.

Synth.pitchbend — Function

pitchbend(chan::Int, signedPB::Int)
pitchbend(chan::Int, signedPB::AbstractFloat)

Sends a 14-bit pitch bend value on the given channel. The sensitivity of this value will depend on the instrument. The AbstractFloat version will rescale (i.e. multiply) by 127 to construct a signed 14-bit value. So values < 1.0 will correspond to microtonal pitch bends.

Synth.midinop — Constant

midinop :: MIDIMsg

A constant MIDIMsg that represents a "no-op" or "nothing to be sent out".

Synth.ismidinop — Function

ismidinop(m::MIDIMsg) :: Bool

Returns true if the given message is a midinop.