Transformers

These operators transform signals in some ways such as mapping ranges (linearmap and expmap), wave shaping (waveshape).

Synth.waveshape — Function

waveshape(f, sig :: Signal)
map(f, sig :: Signal)

Maps a function over the signal. The result is a signal. map and waveshape are aliases for signals.

Synth.linearmap — Function

linearmap(a1 :: Real, a2 :: Real, b1 :: Real, b2 :: Real, s :: Signal)

Maps the input signal from its range to a new range.

Synth.expmap — Function

expmap(a1 :: Real, a2 :: Real, b1 :: Real, b2 :: Real, s :: Signal)

Similar to the job of linearmap but does exponential interpolated mapping. This implies all four values and the signal value must be positive.