Wavetable synthesis

Common synthesis technique used in sampling synths.

Synth.wavetable — Function

wavetable(table :: Vector{Float32}, amp :: Signal, phase :: Signal)

A simple wavetable synth that samples the given table using the given phasor and scales the table by the given amplitude modulator.

The wavetable is a simple array of samples, often in the same sampling rate as the output audio end point, but not necessarily so. The phase signal, which can be generated using phasor, has values in the range $[0,1]$ with $0$ mapping to the start of the wave table and $1$ to the end. To be precise, if the wave table has $N$ samples, then the phase values in the range $[0,1/N)$ will map to the first sample at index $1$ and the sample at index $k$ will be chosen by phase value in the range $[(k-1)/N, k/N)$.

The wavetable processor does 4-point interpolation. This means you can take a small sample table and stretch it out using the interpolation by using a slowly varying phasor. This is how the "pitch" of the sample gets changed during synthesis.

The amplitude modulation is just as with sinosc.

The samples wave table is deemed to be completed when the phasor signal or the amplitude modulator signal are completed.

Synth.maketable — Function

maketable(L :: Int, f)

Utility function to construct a table for use with wavetable. f is passed values in the range [0.0,1.0] to construct the table of the given length L.