For music

Compound musical signals usually have some additional structure that is worthy of special support. At the moment, we support signals with regular beats driven on the falling edge of a sawtooth phasor.

Synth.beats — Function

beats(tempo :: Signal, gen :: BeatGen; phase = 1.0f0, count = -1) :: Beats
beats(tempo :: Real, gen :: BeatGen; phase=1.0, count=-1) :: Beats

Generates beats using genbeat starting with the given phase. The default values are such that the very first sample produced is considered to be the start of a beat. If you need to start it a little later, supply a corresponding value of phase that's < 1.0.

Synth.BeatGen — Type

abstract type BeatGen end

Make a subtype of this type and define the genbeat method on it to use your own custom beat generation algorithm via beats.

Synth.genbeat — Function

genbeat(bg :: BeatGen,
        phase :: AbstractFloat,
        count :: Integer,
        trigger :: Bool,
        dphase :: AbstractFloat,
        dt :: AbstractFloat) :: Union{Val{:end},Nothing,Signal}

Abstract method to be implemented for BeatGen types for use with beats.

phase is (usually) a number in the range [0.0,1.0] that indicates the phase within a beat.

count is the current beat count starting at 0 for the first beat.

trigger is true for the first call of every beat interval.

dphase is the change in phase from previous value. The difference in phase values between this call and the previous.

dt is the usual sampling interval.

This method is called for every sample, letting you add signals during any point within a beat cycle. The return value is expected nothing to indicate nothing to do but continue, Val(:end) to indicate the end of the sequence after which no genbeat call will be made, or a Signal to start playing at the time.