Realtime playback

Methods that provide support for realtime playback as opposed to static file based rendering. The commonly used method is play. To work at a higher level with sample accurate scheduling of sound events, see Gen and the associated gens track, loop and the example ping. For interactivity, see control (for UI to signal) and probe (for signal to UI).

Synth.startaudio — Function

startaudio(callback)

callback is a function that will be called like callback(sample_rate, readqueue, writequeue). sample_rate indicates the output sampling rate. readqueue and writequeue are Channels that accept either an audio buffer as a Vector{Float32} or Val(:done) to indicate completion of the audio process. The callback is expected to take! a buffer from the readqueue, fill it up and submit it to the writequeue using put!.

When the audio generation is done, callback should put!(writequeue, Val(:done)) and break out of its generation loop and return.

It returns a function that can be called (without any arguments) to stop the audio processing thread. Make sure to start julia with a sufficient number of threads for this to work.

Synth.play — Function

play(signal, duration_secs=Inf; blocksize=64)

Plays the given signal on the default audio output in realtime for the given duration or till the signal ends, whichever happens earlier. Be careful about the signal level since in this case the signal can't be globally normalized.

Returns a stop function (like with startaudio and synthesizer) which when called with no arguments will stop playback.

Synth.mic — Function

mic(dev :: AbstractString = "mic"; blocksize :: Int = 128, samplingrate :: Float64 = 48000.0)

Opens input from the default mono mic source. You shouldn't need to fiddle with the blocksize and samplingrate, but they're available for configuration.

A Mic supports intrinsic fanout and therefore you shouldn't need to create multiple mic instances on the same audio interface.

Todo: At some point, support selecting named audio sources and also stereo sources.

Synth.control — Function

control(chan :: Channel{Float32}, dezipper_interval = 0.04; initial = 0.0f0, samplingrate=48000) :: Control
control(dezipper_interval = 0.04; initial = 0.0f0, samplingrate = 48000) :: Control

A "control" is a signal that is driven by values received on a given or created channel. The control will dezipper the value using a first order LPF and send it out as its value. The intention is to be able to bind a UI element that produces a numerical value as a signal that can be patched into the graph.

If c is a Control struct, you can set the value of the control using c[] = 0.5f0.

Close the channel to mark the control signal as "done".

Channels and memory

A control signal uses a channel to receive its values. This raises a question about the amount of memory that'll be consumed by using what looks like a system resource. Julia's channels cost about 416 bytes each, meaning a 1000 channels, which would be a pretty complex scenario to put it mildly, will be well under 1MB. Even if you have 1000 voices with 10 channels controlling each voice, the memory won't be significant (under 5MB) by 2025 standards.

Synth.stop — Function

stop(c :: Control)

Stops the control signal. From the renderer's perspective, the control signal will switch to the "done" state. The control channel will close, causing any further put! calls to raise an exception. If you control the sustain of an adsr using a control signal, then stopping the control will basically end the ADSR envelope by switching it into "release" phase.

Synth.probe — Function

probe(s :: Signal, chan :: Channel{Float32}, interval :: Float64 = 0.04; samplingrate = 48000) :: Probe
probe(s :: Signal, interval :: Float64 = 0.04; samplingrate = 48000) :: Probe

A probe is a "pass through" signal transformer which periodically reports a reading of the signal to a channel. The channel may either be given or a new one can be created by the second variant. Since it is a pass-through, you can insert a probe at any signal point. A probe low-pass-filters the signal before sending it out the channel, so it won't be useful for signals that vary very fast. The default value has it sampling the signal every 40ms.

The channel can then be connected to a UI display widget that shows values as they come in on the channel.

Synth.waveprobe — Function

waveprobe(s :: Signal, chan :: Channel{TimedSamples}, duration :: Float64 = 0.25, interval :: Float64 = 0.04; samplingrate = 48000) :: Probe
waveprobe(s :: Signal, duration :: Float64 = 0.25, interval :: Float64 = 0.04; samplingrate = 48000) :: Probe

A probe is a "pass through" signal transformer which periodically reports a reading of the signal to a channel. The channel may either be given or a new one can be created by the second variant. Since it is a pass-through, you can insert a probe at any signal point. A probe low-pass-filters the signal before sending it out the channel, so it won't be useful for signals that vary very fast. The default value has it sampling the signal every 40ms.

The channel can then be connected to a UI display widget that shows values as they come in on the channel.

STATUS: Needs testing. Also needs a corresponding UI element to test its utility.

Synth.level — Function

level(s::Signal; interval=0.015, refmin=1/(32767*32767), samplingrate=48000)

Computes the smoothed dB level of a signal. The range is from 0 to about 90dB.

refmin is the tiniest sliver of sound intensity that can be registered. The default is set to a value appropriate for 16-bit sampled sound.

interval is the smoothing interval - i.e. the time constant of the first order filter that's applied on the square of the signal.

refmin determines the range since it is the floor of the signal that is 0dB. For the default value, the max ends up around $20\log_{10}(32767) \approx 90.309\text{dB}$. A factor of two change in amplitude is a change in level of around 6.021 dB.

Synth.bus — Function

bus(clk :: Signal) :: Bus

Creates a "bus" which is itself a signal that can be composed with other signals and processors. The bus runs on its own clock and sending Tuple{Float64,Signal} values on the .gchan property will trigger those signals at the given times according to the clock. The scheduling is sample accurate.

A "bus" supports fanout.

bus(tempo_bpm::Real)

Simpler constructor for a fixed tempo bus.

Synth.sched — Function

sched(sch :: Bus{Clk}, t::Union{Beat,Float64}, s::Signal) where {Clk <: Signal}
sched(sch::Bus{Clk}, s::Signal) where {Clk<:Signal}
sched(sch::Bus{Clk}, t::Union{Beat,Float64}, g::Gen) where {Clk<:Signal}
sched(sch::Bus{Clk}, g::Gen) where {Clk<:Signal}

Schedules a signal to start at time t according to the clock of the given bus. In the third variant without a t, the scheduling happens at an ill-specified "now" - which basically means "asap".

Synth.now — Function

now(s::Bus{Clk})::Float64 where {Clk<:Signal}
now(s::Bus{Clk}, b::Type{Beat})::Float64 where {Clk<:Signal}
now(s::Bus{Clk}, b::Beat)::Float64 where {Clk<:Signal}

Returns the bus' "current time". If a beat is passed or asked for, the time is quantized to beats according to the clock's tempo.

Synth.later — Function

later(delay_secs :: Float64, s :: Signal)

Postpones the signal by the given delaysecs. Note that this is not the same as a delay line where there is memory allocated to store some of the samples. The signal is not touched until `delaysecs` has passed, and the time value that the signal ends up seeing also does not span the period up to the delay.

Gens

A "Gen" has a method to implement which has the signature -

proc(::Gen, ::Bus, t::Float64) -> Tuple{Float64,Gen}

The proc method is expected to schedule any signals to the bus using sched(::Bus,t,::Signal) and return info about the next gen to trigger later. The t argument to proc is "clock time" and not real time - i.e. it is determined by the clock which the bus runs.

Missing docstring.

Missing docstring for Synth.Gen. Check Documenter's build log for details.

Synth.seq — Function

seq(ga :: GA, gb :: GB) :: Gen where {GA <: Gen, GB <: Gen}

Sequences the given two gens to occur one after the other. When the first gen ga results in a Cont, it will switch to gb and after that will continue onwards.

Synth.track — Function

track(gs :: AbstractVector) :: Gen

A "track" is a heterogeneous sequence of Gens. When each gen finishes, its proc call will produce a Cont which indicates it is time to switch to the next Gen in the track. This is like a generalization of seq except that seq has types known at construction time.

Synth.durn — Function

durn(d :: Real, gen :: Gen) :: Gen

Forces the duration of the given gen to be the given d, even if it was something else. Possibly useful with chord.

Synth.chord — Function

chord(gens :: AbstractVector) :: Gen

A Gen which runs all the given heterogeneous gens in lock step. The duration of the chord is given by the maximum of the durations of all the gens every time they run. Whenever a gen produces a Cont or Stop, it is taken out of the set of gens and the other continue on.

Broadly, chord is useful when all the gens have the same duration - very much like a chord in music. If you're thinking of whether you should use this to start off multiple musical processes in parallel with each determining its own duration, you probably want par.

Synth.par — Function

par(gens :: AbstractVector) :: Par

For parallel composition of gens given as a vector. The Par itself has effectively zero duration and when sequenced as part of a track will result in the following gen being immediately processed without delay. The individual gens part of each of the "threads" will continue to be processed by the bus over time until they finish.

Synth.loop — Function

loop(n :: Int, g :: Gen) :: Gen

Will loop the given gen n times before ending.

Synth.dyn — Function

dyn(fn :: Function, n :: Int, i :: Int) :: Gen

A "dyn" gen calls the function with each i to determine the gen that should be performed. The function will be called with two arguments i and n where i will range from 1 to n.

Synth.ping — Function

ping(pitch :: Real, dur :: Real, vel :: Real = 0.5f0, decay :: Real = dur) :: Gen
ping(pitch :: AbstractVector{R}, dur :: Real, vel :: Real = 0.5f0, decay :: Real = dur) :: Gen where {R <: Real}
ping(pitch :: AbstractVector{R}, dur :: AbstractVector{RD}, vel :: Real = 0.5f0, decay :: Real = dur) :: Gen where {R <: Real, RD <: Real}
ping(pch :: PitchChord, dur :: Real, vel :: Real = 0.5f0, decay :: Real = dur) :: Gen

A "ping" is a simple decaying sine tone for illustrating how to create a Gen. If you use the array versions, they're made into corresponding tracks for ease of use. So ping([60,67,72], 0.5) would be a three-note track. If the duration is also an array, it will be cycled through for each of the pitch values. So the number of pings is determined only by the number of pitches given.

Synth.ch — Function

ch(pitches :: AbstractVector{R}) :: PitchChord where {R <: Real}

Represents a set of chorded pitch values. Usable with ping.